pyproject.toml
[project]
name = "google-lens-to-mercari"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "google-generativeai>=0.8.6",
    "google-search-results>=2.4.2",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "playwright>=1.57.0",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-playwright>=0.7.2",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "requests>=2.32.5",
    "uvicorn[standard]>=0.40.0",
]

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]
pythonpath = [
  "."
]
